requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic_settings
pydantic[email]
python-jose
passlib
alembic